Fix crash in function CGImageDestinationFinalize of ImageIO framework when generating thumbnails.
- The crash vImage: _ERROR_Buffer_Read__Too_Small_For_Arguments_To_vImage__CheckBacktrace + 56 was happening when generating thumbnails. The calculated size for screenshots were 200 x (2048/1536) = 266 => 266 / (2048/1536) = 199 => thumbnails of 199x199 , therefore producing a crash when expecting 200 x 200 images. Rounding to the next higher integer with ceil, the crash is fixed.

+++ include/mega/gfx/GfxProcCG.h
@@ -32,8 +32,8 @@ class MEGA_API GfxProcCG : public mega::GfxProc
     CGImageSourceRef imageSource;
     CFDictionaryRef imageParams;
     CFMutableDictionaryRef thumbnailParams;
-    CGImageRef createThumbnailWithMaxSize(int size);
-    int maxSizeForThumbnail(const int rw, const int rh);
+    CGImageRef createThumbnailWithMaxSize(double size);
+    double maxSizeForThumbnail(const int rw, const int rh);
 private: // mega::GfxProc implementations
     const char* supportedformats();
     bool readbitmap(mega::FileAccess*, mega::string*, int);
